# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DiffResponse": {
        "properties": {
          "baseline": {
            "$ref": "#/components/schemas/TimeWindow"
          },
          "current": {
            "$ref": "#/components/schemas/TimeWindow"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/DiffEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "filters": {
            "$ref": "#/components/schemas/EffectiveFilters"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "include_unchanged": {
            "title": "Include Unchanged",
            "type": "boolean"
          },
          "summary": {
            "$ref": "#/components/schemas/DiffSummary"
          },
          "threshold_percent": {
            "title": "Threshold Percent",
            "type": "number"
          }
        },
        "required": [
          "generated_at",
          "baseline",
          "current",
          "threshold_percent",
          "filters",
          "include_unchanged",
          "edges",
          "summary"
        ],
        "title": "DiffResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Stable error envelope. Never carries a stack trace, DSN, or credential (ADR-004 D-4.6).",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "error",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Runtime service topology for Kubernetes, collected with eBPF.\n\nThis document is the contract between the backend, the Go node agent, and the React frontend.\nIt is generated from the FastAPI application and committed as `contracts/openapi.json`; CI fails\nif the committed copy drifts. Never hand-edit it.\n\nIdentity, edge keys, status codes, and determinism rules are specified in `contracts/ids.md`.",
    "title": "Kubernetes Runtime Topology Visualizer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/diff": {
      "get": {
        "description": "Deterministic comparison of two periods.\n\nA period with no observation contributes zero, not null. Percentage delta is undefined when\nthe baseline is zero, and the reason field says so.",
        "operationId": "get_diff_api_v1_diff_get",
        "parameters": [
          {
            "in": "query",
            "name": "baseline_from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Baseline From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "baseline_to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Baseline To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "current_from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Current From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "current_to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Current To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Namespace"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "Service",
                    "Deployment",
                    "StatefulSet",
                    "DaemonSet",
                    "Job",
                    "Pod"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "include_external",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include External",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_unresolved",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Unresolved",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_unchanged",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Unchanged",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Diff",
        "tags": [
          "topology"
        ]
      }
    },
    "/api/v1/edges/{edge_id}": {
      "get": {
        "description": "Edge detail. `edge_id` is `source|target|protocol|port` and must be URL-encoded.",
        "operationId": "get_edge_api_v1_edges__edge_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "edge_id",
            "required": true,
            "schema": {
              "title": "Edge Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "1m",
                    "5m",
                    "15m",
                    "1h",
                    "6h",
                    "24h"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Window"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Edge",
        "tags": [
          "topology"
        ]
      }
    },
    "/api/v1/graph": {
      "get": {
        "description": "Topology for one window. Nodes are derived only from edges inside that window.\n\nSupply exactly one of `window` or the `from`/`to` pair.",
        "operationId": "get_graph_api_v1_graph_get",
        "parameters": [
          {
            "description": "Preset window. Mutually exclusive with from/to.",
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "1m",
                    "5m",
                    "15m",
                    "1h",
                    "6h",
                    "24h"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Preset window. Mutually exclusive with from/to.",
              "title": "Window"
            }
          },
          {
            "description": "RFC 3339 UTC, inclusive.",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "RFC 3339 UTC, inclusive.",
              "title": "From"
            }
          },
          {
            "description": "RFC 3339 UTC, exclusive.",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "RFC 3339 UTC, exclusive.",
              "title": "To"
            }
          },
          {
            "description": "Repeatable.",
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeatable.",
              "title": "Namespace"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "Service",
                    "Deployment",
                    "StatefulSet",
                    "DaemonSet",
                    "Job",
                    "Pod"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "description": "Case-insensitive substring match on label.",
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring match on label.",
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "include_external",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include External",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_unresolved",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Unresolved",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Graph",
        "tags": [
          "topology"
        ]
      }
    },
    "/api/v1/ingest/batches": {
      "post": {
        "description": "Idempotent ingestion. Re-posting a `batch_id` returns 200 and changes nothing.\n\nVersion checking happens here rather than in the model so that an unsupported version yields\n400 rather than Pydantic's 422 (contracts/ids.md §8).",
        "operationId": "ingest_batch_api_v1_ingest_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResult"
                }
              }
            },
            "description": "batch_id already ingested; no state change"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported schema_version"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Body or edge count over limit"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed batch"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Storage unavailable; retry"
          }
        },
        "summary": "Ingest Batch",
        "tags": [
          "topology"
        ]
      }
    },
    "/api/v1/namespaces": {
      "get": {
        "description": "Namespaces observed in the window, for populating the filter panel.",
        "operationId": "list_namespaces_api_v1_namespaces_get",
        "parameters": [
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "1m",
                    "5m",
                    "15m",
                    "1h",
                    "6h",
                    "24h"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Window"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Namespaces",
        "tags": [
          "topology"
        ]
      }
    },
    "/api/v1/nodes/{node_id}": {
      "get": {
        "description": "Node detail. `node_id` contains ':' separators and must be URL-encoded by the client.",
        "operationId": "get_node_api_v1_nodes__node_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "1m",
                    "5m",
                    "15m",
                    "1h",
                    "6h",
                    "24h"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Window"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Node",
        "tags": [
          "topology"
        ]
      }
    },
...
  }
}